Update components based on given structure
What:
- For every component in the components folder I created a dedicated folder
- Each component folder has a [component.js] and a styles.js file
- Cleaned up some code
- Added developer checklist

Why:
- With growing number of components it will be easier to search for individual components rather than filtering through index.js in search results
- I moved the styles to its own styles.js file because it makes it easier to look at the react code isolated from it's styles
- My IDE indicated there were many eslint errors. Hence the small clean ups
- Checklist may provide some measures to ensure we are adding quality code.
