+++ src/components/ui/GetInvolved/GetInvolved.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import Container from '../../Container';
+import StyledGetInvolved from './styles';
+
+const GetInvolved = () => (
+  <Container>
+    <StyledGetInvolved>
+      <div className="image-container">
+        <div className="image-text-container">
+          <div className="image-text">
+            <h3>#BLACKLIVESMATTER</h3>
+            <h3>How to get involved</h3>
+          </div>
+        </div>
+      </div>
+      <div className="details">
+        <h1>Delayed justice is injustice</h1>
+      </div>
+    </StyledGetInvolved>
+  </Container>
+);
+
+export default GetInvolved;
