+++ src/components/ui/MediaList/styles.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+const StyledMediaList = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
+  justify-content: space-evenly;
+
+  h2 {
+    font-family: Karla;
+    font-style: normal;
+    font-weight: bold;
+    width: 100%;
+    margin: 1rem 0 0 4rem;
+    text-transform: uppercase;
+  }
+`;
+
+export default StyledMediaList;
